component restructuring
